summarizer-app/requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
langchain-ollama
